Added lib module to detect if a page can be deleted
